rearranging/renaming classes adding test cases to TransactionManagerTest

+++ src/main/java/entities/Item.java
@@ -16,7 +16,7 @@ public class Item {
     private String name;
 
     @Column
-    private long price;
+    private int price;
 
     public String getName() {
         return name;
@@ -26,11 +26,11 @@ public class Item {
         this.name = name;
     }
 
-    public long getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(long price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 }
